% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_popover.R
\name{extract_createTooltipOrPopoverOnUI}
\alias{extract_createTooltipOrPopoverOnUI}
\title{Tooltips and Popovers}
\usage{
extract_createTooltipOrPopoverOnUI(id, type, options)
}
\description{
Tooltips and Popovers allow you to add additional information about controls
or outputs without cluttering up your user interface. You can add a tooltip to
a button that displays on hover and better explains what the button will do, or
you could add a popover to an output providing further analysis of that output.
}
\details{
You can create tooltips and popovers from either the UI script or within the
Server logic. \code{\link{bsTooltip}} and \code{\link{extract_bsPopover}} are used in
the UI, and \code{\link{addTooltip}} and \code{\link{addPopover}} are used in
the Server logic. \code{\link{tipify}} and \code{\link{popify}} can be used
within the UI or from within a \code{\link{renderUI}} in the Server logic. They
also have the added advantage of not requiring that the UI element have an ID
attribute.
}
\section{Components}{

There are eight functions in the Tooltips and Popovers family:
 \describe{
   \item{\code{\link{bsTooltip}}}{Used in the UI to add a tooltip to an element
   in your UI.}
   \item{\code{\link{extract_bsPopover}}}{Used in the UI to add a popover to an element
   in your UI.}
   \item{\code{\link{tipify}}}{Wrap any UI element in \code{tipify} to add a
   tooltip to the wrapped element. Preferred for elemented created with
   \code{\link{renderUI}}.}
   \item{\code{\link{popify}}}{Wrap any UI element in \code{popify} to add a
   popover to the wrapped element. Preferred for elements created with
   \code{\link{renderUI}}.}
   \item{\code{\link{addTooltip}}}{Used in the Server logic to add a tooltip
   to an element in your UI.}
   \item{\code{\link{addPopover}}}{Used in the Server logic to add a popover
   to an element in your UI.}
   \item{\code{\link{removeTooltip}}}{Used in the Server logic to remove a
   tooltip from an element in your UI.}
   \item{\code{\link{removePopover}}}{Used in the Server logic to remove a
   popover from an element in your UI.}
 }
}

